YARN-11511 Improve TestRMWebServices test config and data
delete dups
