+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/AbstractParentQueue.java
@@ -548,6 +548,11 @@ public abstract class AbstractParentQueue extends AbstractCSQueue {
    * false otherwise
    */
   public boolean isEligibleForAutoQueueCreation() {
+    CapacitySchedulerConfiguration configuration = queueContext.getConfiguration();
+    System.err.println(configuration);
+    boolean autoQueueCreationV2Enabled =
+        queueContext.getConfiguration().isAutoQueueCreationV2Enabled(getQueuePath());
+    System.err.println(autoQueueCreationV2Enabled);
     return isDynamicQueue() || queueContext.getConfiguration().
         isAutoQueueCreationV2Enabled(getQueuePath());
   }
