+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/java/org/apache/hadoop/yarn/server/resourcemanager/webapp/TestRMWebServicesCapacitySchedDynamicConfig.java
@@ -18,165 +18,154 @@
 
 package org.apache.hadoop.yarn.server.resourcemanager.webapp;
 
+import java.io.IOException;
+import javax.ws.rs.core.MediaType;
+
 import com.google.inject.Guice;
 import com.sun.jersey.api.client.ClientResponse;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.ws.rs.core.MediaType;
+import org.junit.Test;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.yarn.conf.YarnConfiguration;
-import org.apache.hadoop.yarn.exceptions.YarnException;
+import org.apache.hadoop.yarn.server.resourcemanager.MockNM;
 import org.apache.hadoop.yarn.server.resourcemanager.MockRM;
-import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.AutoCreatedQueueTemplate;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.ResourceScheduler;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacityScheduler;
-import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerQueueManager;
 import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.QueuePath;
 import org.apache.hadoop.yarn.webapp.GuiceServletConfig;
 import org.apache.hadoop.yarn.webapp.JerseyTestBase;
-import org.junit.Test;
 
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerTestUtilities.GB;
 import static org.apache.hadoop.yarn.server.resourcemanager.webapp.TestRMWebServicesCapacitySched.assertJsonResponse;
-import static org.apache.hadoop.yarn.server.resourcemanager.webapp.TestRMWebServicesCapacitySched.createMockRM;
 import static org.apache.hadoop.yarn.server.resourcemanager.webapp.TestRMWebServicesCapacitySched.createWebAppDescriptor;
+import static org.assertj.core.api.Assertions.fail;
 
-/**
- * @deprecated Please use the {@link TestRMWebServicesCapacitySchedDynamicConfigV2} for new tests.
- */
-@Deprecated
-public class TestRMWebServicesCapacitySchedDynamicConfig extends
-    JerseyTestBase {
+public class TestRMWebServicesCapacitySchedDynamicConfig extends JerseyTestBase {
+
+  private static final String EXPECTED_FILE_TMPL = "webapp/dynamic-%s-%s.json";
   private MockRM rm;
-
-  private CapacitySchedulerQueueManager autoQueueHandler;
 
   public TestRMWebServicesCapacitySchedDynamicConfig() {
     super(createWebAppDescriptor());
   }
 
-  @Test // mar van ilyen
-  public void testSchedulerResponsePercentageMode()
-      throws Exception {
-    Configuration config = createPercentageConfig();
+  private static final String[] SETUP = new String[] {
+      "yarn.scheduler.capacity.root.queues = default, test1, test2",
+      "yarn.scheduler.capacity.root.test1.queues = test1_1, test1_2, test1_3"
+  };
 
-    initResourceManager(config);
+  /*
+   *                                         EffectiveMin (32GB 32VCores)     AbsoluteCapacity
+   *     root.default              4/32      [memory=4096,  vcores=4]       12.5%
+   *     root.test_1              16/32      [memory=16384, vcores=16]
+   *     root.test_2              12/32      [memory=12288, vcores=12]      37.5%
+   *     root.test_1.test_1_1      2/16      [memory=2048,  vcores=2]       6.25%
+   *     root.test_1.test_1_2      2/16      [memory=2048,  vcores=2]       6.25%
+   *     root.test_1.test_1_3     12/16      [memory=12288, vcores=12]      37.5%
+   */
 
-    /*
-     * mode: percentage
-     * autoCreationEligibility: off
-     * weight: -1, normalizedWeight: 0
-     * root.queueType: parent, others.queueType: leaf
-     */
-    assertJsonResponse(sendRequest(), "webapp/scheduler-response-PercentageMode.json");
+  @Test
+  public void testPercentageMode() throws Exception {
+    runTest("testPercentageMode",
+        "yarn.scheduler.capacity.root.default.capacity = 12.5",
+        "yarn.scheduler.capacity.root.test1.capacity =                 50",
+        "yarn.scheduler.capacity.root.test2.capacity =                 37.5",
+        "yarn.scheduler.capacity.root.test1.test1_1.capacity =         12.5",
+        "yarn.scheduler.capacity.root.test1.test1_2.capacity =         12.5",
+        "yarn.scheduler.capacity.root.test1.test1_3.capacity =         75"
+    );
+  }
+  @Test
+  public void testAbsoluteMode() throws Exception {
+    runTest("testAbsoluteMode",
+        "yarn.scheduler.capacity.root.default.capacity = [memory=4096,vcores=4]",
+        "yarn.scheduler.capacity.root.test1.capacity = [memory=16384,vcores=16]",
+        "yarn.scheduler.capacity.root.test2.capacity = [memory=12288,vcores=12]",
+        "yarn.scheduler.capacity.root.test1.test1_1.capacity = [memory=2048,vcores=2]",
+        "yarn.scheduler.capacity.root.test1.test1_2.capacity = [memory=2048,vcores=2]",
+        "yarn.scheduler.capacity.root.test1.test1_3.capacity = [memory=12288,vcores=12]"
+    );
   }
 
-  @Test // uj test kell
-  public void testSchedulerResponsePercentageModeLegacyAutoCreation()
-      throws Exception {
-    Configuration config = createPercentageConfigLegacyAutoCreation();
-
-    initResourceManager(config);
-
-    /*
-     * mode: percentage
-     * managedtest2.autoCreationEligibility: legacy, others.autoCreationEligibility: off
-     * weight: -1, normalizedWeight: 0
-     * root.queueType: parent, others.queueType: leaf
-     */
-    assertJsonResponse(sendRequest(),
-        "webapp/scheduler-response-PercentageModeLegacyAutoCreation.json");
+  @Test
+  public void testWeightMode() throws Exception {
+    String[] capacity = new String[] {
+        "yarn.scheduler.capacity.root.default.capacity =       4w",
+        "yarn.scheduler.capacity.root.test1.capacity =         16w",
+        "yarn.scheduler.capacity.root.test2.capacity =         12w",
+        "yarn.scheduler.capacity.root.test1.test1_1.capacity = 2w",
+        "yarn.scheduler.capacity.root.test1.test1_2.capacity = 2w",
+        "yarn.scheduler.capacity.root.test1.test1_3.capacity = 12w"
+    };
+    runTest("testWeightMode", capacity);
+    runAQCTest("testWeightMode", capacity);
   }
 
-  @Test //uj test kell
-  public void testSchedulerResponseAbsoluteModeLegacyAutoCreation()
-      throws Exception {
-    Configuration config = createAbsoluteConfigLegacyAutoCreation();
-
-    initResourceManager(config);
-    initAutoQueueHandler(8192);
-    createQueue("root.managed.queue1");
-
-    assertJsonResponse(sendRequest(),
-        "webapp/scheduler-response-AbsoluteModeLegacyAutoCreation.json");
+  private void runTest(
+      String name,
+      String... capacitySetup
+  ) throws Exception {
+    Configuration config = new Configuration();
+    enrichConfig(config, SETUP);
+    enrichConfig(config, capacitySetup);
+    createRM(config);
+    MockNM nm1 = rm.registerNode("h1:1234", 8192, 8);
+    MockNM nm2 = rm.registerNode("h2:1234", 8192, 8);
+    reinitialize(config);
+    assertJsonResponse(sendRequest(), String.format(EXPECTED_FILE_TMPL, name, 16));
+    MockNM nm3 = rm.registerNode("h3:1234", 8192, 8);
+    MockNM nm4 = rm.registerNode("h4:1234", 8192, 8);
+    reinitialize(config);
+    assertJsonResponse(sendRequest(), String.format(EXPECTED_FILE_TMPL, name, 32));
+    rm.unRegisterNode(nm1);
+    rm.unRegisterNode(nm4);
+    reinitialize(config);
+    assertJsonResponse(sendRequest(), String.format(EXPECTED_FILE_TMPL, name, 16));
   }
 
-  @Test // van ilyen
-  public void testSchedulerResponseAbsoluteMode()
-      throws Exception {
-    Configuration config = createAbsoluteConfig();
-
-    initResourceManager(config);
-    rm.registerNode("h1:1234", 102400, 100);
-
-    /*
-     * mode: absolute
-     * autoCreationEligibility: off
-     * weight: -1, normalizedWeight: 0
-     * root.queueType: parent, others.queueType: leaf
-     */
-    assertJsonResponse(sendRequest(), "webapp/scheduler-response-AbsoluteMode.json");
+  private void runAQCTest(
+      String name,
+      String... capacitySetup
+  ) throws Exception {
+    Configuration config = new Configuration();
+    enrichConfig(config, SETUP);
+    enrichConfig(config, capacitySetup);
+    setupAQC(config, "yarn.scheduler.capacity.root.test1.");
+    setupAQC(config, "yarn.scheduler.capacity.root.test1.test1_1.");
+    createRM(config);
+    createAQC("test1");
+    createAQC("test1.test1_1");
+    rm.registerNode("h1:1234", 32768, 32);
+    reinitialize(config);
+    assertJsonResponse(sendRequest(), String.format(EXPECTED_FILE_TMPL, name, "aqc"));
   }
 
-  @Test // ilyen is van
-  public void testSchedulerResponseWeightMode()
-      throws Exception {
-    Configuration config = createWeightConfig();
-
-    initResourceManager(config);
-
-    /*
-     * mode: weight
-     * autoCreationEligibility: off
-     *                   root   default  test1  test2
-     * weight:            1       10       4     6
-     * normalizedWeight:  1       0.5      0.2   0.3
-     * root.queueType: parent, others.queueType: leaf
-     */
-    assertJsonResponse(sendRequest(), "webapp/scheduler-response-WeightMode.json");
+  private void setupAQC(Configuration config, String queue) {
+    config.set(queue + "auto-queue-creation-v2.enabled", "true");
+    config.set(queue + "auto-queue-creation-v2.maximum-queue-depth", "10");
+    config.set(queue + "auto-queue-creation-v2.parent-template.acl_submit_applications", "u1");
+    config.set(queue + "auto-queue-creation-v2.parent-template.acl_administer_queue", "u2");
+    config.set(queue + "autoParent1.auto-queue-creation-v2.leaf-template.acl_submit_applications", "u3");
+    config.set(queue + "autoParent1.auto-queue-creation-v2.leaf-template.acl_administer_queue", "u4");
+    config.set(queue + "*.auto-queue-creation-v2.leaf-template.acl_submit_applications", "u5");
+    config.set(queue + "*.auto-queue-creation-v2.leaf-template.acl_administer_queue", "u6");
+    config.set(queue + "parent.*.auto-queue-creation-v2.leaf-template.acl_submit_applications", "u7");
+    config.set(queue + "parent.*.auto-queue-creation-v2.leaf-template.acl_administer_queue", "u8");
+    config.set(queue + "autoParent1.auto-queue-creation-v2.template.maximum-applications", "300");
   }
-
-  @Test // ilyen is van
-  public void testSchedulerResponseWeightModeWithAutoCreatedQueues()
-      throws Exception {
-    Configuration config = createWeightConfigWithAutoQueueCreationEnabled();
-    config.setInt(CapacitySchedulerConfiguration
-        .getQueuePrefix("root.autoParent1") +
-        AutoCreatedQueueTemplate.AUTO_QUEUE_TEMPLATE_PREFIX +
-        "maximum-applications", 300);
-
-    initResourceManager(config);
-    initAutoQueueHandler(1200);
-
-    // same as webapp/scheduler-response-WeightMode.json, but with effective resources filled in
-    assertJsonResponse(sendRequest(),
-        "webapp/scheduler-response-WeightModeWithAutoCreatedQueues-Before.json");
-
-    //Now create some auto created queues
-    createQueue("root.auto1");
-    createQueue("root.auto2");
-    createQueue("root.auto3");
-    createQueue("root.autoParent1.auto4");
-    createQueue("root.autoParent2.auto5");
-    createQueue("root.parent.autoParent3.auto6");
-
-    assertJsonResponse(sendRequest(),
-        "webapp/scheduler-response-WeightModeWithAutoCreatedQueues-After.json");
-  }
-
-  private void initAutoQueueHandler(int nodeMemory) throws Exception {
-    CapacityScheduler cs = (CapacityScheduler) rm.getResourceScheduler();
-    autoQueueHandler = cs.getCapacitySchedulerQueueManager();
-    rm.registerNode("h1:1234", nodeMemory, 32); // label = x
-  }
-
-  private void createQueue(String queuePath) throws YarnException,
-      IOException {
-    autoQueueHandler.createQueue(new QueuePath(queuePath));
+  private void createAQC(String queue) {
+    try {
+      CapacityScheduler cs = (CapacityScheduler) rm.getResourceScheduler();
+      CapacitySchedulerQueueManager autoQueueHandler = cs.getCapacitySchedulerQueueManager();
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".auto1"));
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".auto2"));
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".auto3"));
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".autoParent1.auto4"));
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".autoParent2.auto5"));
+      autoQueueHandler.createQueue(new QueuePath("root." + queue + ".parent.autoParent3.auto6"));
+    } catch (Throwable e) {
+      fail("Can not auto create queues under " + queue, e);
+    }
   }
 
   private ClientResponse sendRequest() {
@@ -185,142 +174,31 @@ public class TestRMWebServicesCapacitySchedDynamicConfig extends
         .get(ClientResponse.class);
   }
 
-  private Configuration createPercentageConfig() {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("yarn.scheduler.capacity.root.queues", "default, test1, test2");
-    conf.put("yarn.scheduler.capacity.root.test1.capacity", "50");
-    conf.put("yarn.scheduler.capacity.root.test2.capacity", "50");
-    conf.put("yarn.scheduler.capacity.root.test1.maximum-capacity", "100");
-    conf.put("yarn.scheduler.capacity.root.test1.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.test2.state", "RUNNING");
-    return createConfiguration(conf);
+  private void reinitialize(Configuration conf) throws IOException {
+    CapacityScheduler cs = (CapacityScheduler) rm.getResourceScheduler();
+    cs.reinitialize(conf, rm.getRMContext(), true);
   }
 
-  private Configuration createPercentageConfigLegacyAutoCreation() {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("yarn.scheduler.capacity.root.queues", "default, test1, " +
-        "managedtest2");
-    conf.put("yarn.scheduler.capacity.root.test1.capacity", "50");
-    conf.put("yarn.scheduler.capacity.root.managedtest2.capacity", "50");
-    conf.put("yarn.scheduler.capacity.root.test1.maximum-capacity", "100");
-    conf.put("yarn.scheduler.capacity.root.test1.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.managedtest2.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.managedtest2." +
-        "auto-create-child-queue.enabled", "true");
-    return createConfiguration(conf);
-  }
-
-  private Configuration createAbsoluteConfigLegacyAutoCreation() {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("yarn.scheduler.capacity.root.queues", "default, managed");
-    conf.put("yarn.scheduler.capacity.root.default.state", "STOPPED");
-    conf.put("yarn.scheduler.capacity.root.default.capacity", "[memory=0,vcores=0]");
-    conf.put("yarn.scheduler.capacity.root.managed.capacity", "[memory=4096,vcores=4]");
-    conf.put("yarn.scheduler.capacity.root.managed.leaf-queue-template.capacity",
-        "[memory=2048,vcores=2]");
-    conf.put("yarn.scheduler.capacity.root.managed.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.managed." +
-        "auto-create-child-queue.enabled", "true");
-    conf.put("yarn.scheduler.capacity.root.managed.leaf-queue-template.acl_submit_applications",
-        "user");
-    conf.put("yarn.scheduler.capacity.root.managed.leaf-queue-template.acl_administer_queue",
-        "admin");
-    return createConfiguration(conf);
-  }
-
-  private Configuration createAbsoluteConfig() {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("yarn.scheduler.capacity.root.queues", "default, test1, test2");
-    conf.put("yarn.scheduler.capacity.root.capacity",
-        "[memory=6136,vcores=30]");
-    conf.put("yarn.scheduler.capacity.root.default.capacity",
-        "[memory=3064,vcores=15]");
-    conf.put("yarn.scheduler.capacity.root.test1.capacity",
-        "[memory=2048,vcores=10]");
-    conf.put("yarn.scheduler.capacity.root.test2.capacity",
-        "[memory=1024,vcores=5]");
-    conf.put("yarn.scheduler.capacity.root.test1.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.test2.state", "RUNNING");
-    return createConfiguration(conf);
-  }
-
-  private Configuration createWeightConfig() {
-    return createWeightConfigInternal(false);
-  }
-
-  private Configuration createWeightConfigWithAutoQueueCreationEnabled() {
-    return createWeightConfigInternal(true);
-  }
-
-  private Configuration createWeightConfigInternal(boolean enableAqc) {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("yarn.scheduler.capacity.root.queues", "default, test1, test2, parent");
-    conf.put("yarn.scheduler.capacity.root.capacity", "1w");
-    conf.put("yarn.scheduler.capacity.root.default.capacity", "10w");
-    conf.put("yarn.scheduler.capacity.root.test1.capacity", "5w");
-    conf.put("yarn.scheduler.capacity.root.test2.capacity", "10w");
-    conf.put("yarn.scheduler.capacity.root.parent.capacity", "20w");
-    conf.put("yarn.scheduler.capacity.root.test1.state", "RUNNING");
-    conf.put("yarn.scheduler.capacity.root.test2.state", "RUNNING");
-
-    if (enableAqc) {
-      final String root = "yarn.scheduler.capacity.root.";
-      conf.put(root +  "auto-queue-creation-v2.enabled", "true");
-
-      conf.put(root + "auto-queue-creation-v2.parent-template.acl_submit_applications",
-          "parentUser1");
-      conf.put(root + "auto-queue-creation-v2.parent-template.acl_administer_queue",
-          "parentAdmin1");
-
-      conf.put(root + "autoParent1.auto-queue-creation-v2.leaf-template.acl_submit_applications",
-          "user1");
-      conf.put(root + "autoParent1.auto-queue-creation-v2.leaf-template.acl_administer_queue",
-          "admin1");
-
-      conf.put(root + "*.auto-queue-creation-v2.leaf-template.acl_submit_applications",
-          "wildUser1");
-      conf.put(root + "*.auto-queue-creation-v2.leaf-template.acl_administer_queue",
-          "wildAdmin1");
-
-
-      conf.put(root + "parent.auto-queue-creation-v2.enabled", "true");
-      conf.put(root + "parent.auto-queue-creation-v2.parent-template.acl_submit_applications",
-          "parentUser2");
-      conf.put(root + "parent.auto-queue-creation-v2.parent-template.acl_administer_queue",
-          "parentAdmin2");
-
-      conf.put(root + "parent.*.auto-queue-creation-v2.leaf-template.acl_submit_applications",
-          "wildUser2");
-      conf.put(root + "parent.*.auto-queue-creation-v2.leaf-template.acl_administer_queue",
-          "wildAdmin2");
-    }
-    return createConfiguration(conf);
-  }
-
-  private Configuration createConfiguration(
-      Map<String, String> configs) {
-    Configuration config = new Configuration();
-
-    for (Map.Entry<String, String> entry : configs.entrySet()) {
-      config.set(entry.getKey(), entry.getValue());
-    }
-
+  private void createRM(Configuration config) throws IOException {
+    config.setClass(YarnConfiguration.RM_SCHEDULER,
+        CapacityScheduler.class, ResourceScheduler.class);
+    config.set(YarnConfiguration.RM_PLACEMENT_CONSTRAINTS_HANDLER,
+        YarnConfiguration.SCHEDULER_RM_PLACEMENT_CONSTRAINTS_HANDLER);
     config.set(YarnConfiguration.SCHEDULER_CONFIGURATION_STORE_CLASS,
         YarnConfiguration.MEMORY_CONFIGURATION_STORE);
-
-    return config;
+    rm = new MockRM(config);
+    GuiceServletConfig.setInjector(Guice.createInjector(new TestRMWebServicesCapacitySched.WebServletModule(rm)));
+    rm.start();
+    reinitialize(config);
   }
 
-  private void initResourceManager(Configuration conf) throws IOException {
-    rm = createMockRM(new CapacitySchedulerConfiguration(conf));
-    GuiceServletConfig.setInjector(
-        Guice.createInjector(new TestRMWebServicesCapacitySched.WebServletModule(rm)));
-    rm.start();
-    //Need to call reinitialize as
-    //MutableCSConfigurationProvider with InMemoryConfigurationStore
-    //somehow does not load the queues properly and falls back to default config.
-    //Therefore CS will think there's only the default queue there.
-    ((CapacityScheduler) rm.getResourceScheduler()).reinitialize(conf,
-        rm.getRMContext(), true);
+  private void enrichConfig(Configuration config, String[] newValues) {
+    for (String newValue : newValues) {
+      int index = newValue.indexOf("=");
+      config.set(
+          newValue.substring(0, index).trim(),
+          newValue.substring(index + 1).trim()
+      );
+    }
   }
 }
